Feature/add weapons skills

+++ Services/CharacterService/CharacterService.cs
@@ -1,23 +1,35 @@
+using System.Security.Claims;
 using AutoMapper;
 using dotnet_rpg.Data;
+using dotnet_rpg.Dtos;
 using dotnet_rpg.Dtos.Character;
-using dotnet_rpg.Models;
 using Microsoft.EntityFrameworkCore;
+
+#nullable disable
 
 namespace dotnet_rpg.Services.CharacterService
 {
-#nullable disable
-
     public class CharacterService : ICharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(
+            IMapper mapper,
+            DataContext context,
+            IHttpContextAccessor httpContextAccessor
+        )
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() =>
+            int.Parse(
+                _httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier)
+            );
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(
             AddCharacterDto newCharacter
@@ -25,9 +37,11 @@ namespace dotnet_rpg.Services.CharacterService
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
             serviceResponse.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUserId())
                 .Select(c => _mapper.Map<GetCharacterDto>(c))
                 .ToListAsync();
             return serviceResponse;
@@ -40,12 +54,23 @@ namespace dotnet_rpg.Services.CharacterService
 
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = _context.Characters
-                    .Select(c => _mapper.Map<GetCharacterDto>(c))
-                    .ToList();
+                Character character = await _context.Characters.FirstOrDefaultAsync(
+                    c => c.Id == id && c.User.Id == GetUserId()
+                );
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c))
+                        .ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Could not find character.";
+                }
             }
             catch (Exception ex)
             {
@@ -56,22 +81,27 @@ namespace dotnet_rpg.Services.CharacterService
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
-            var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
+            var response = new ServiceResponse<List<GetCharacterDto>>();
             var dbCharacters = await _context.Characters
-                .Where(c => c.User.Id == userId)
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .Where(c => c.User.Id == GetUserId())
                 .ToListAsync();
-            serviceResponse.Data = dbCharacters
-                .Select(c => _mapper.Map<GetCharacterDto>(c))
-                .ToList();
-            return serviceResponse;
+            response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            return response;
         }
 
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .FirstOrDefaultAsync(
+                c => c.Id == id && c.User.Id == GetUserId()
+            );
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -85,36 +115,43 @@ namespace dotnet_rpg.Services.CharacterService
 
             try
             {
-                Character character = await _context.Characters.FirstOrDefaultAsync(
-                    c => c.Id == updatedCharacter.Id
-                );
+                Character character = await _context.Characters
+                    .Include(c => c.User)
+                    .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                if (character.User.Id == GetUserId())
+                {
+                    character.Name =
+                        updatedCharacter.Name == null ? character.Name : updatedCharacter.Name;
+                    character.HitPoints =
+                        updatedCharacter.HitPoints.ToString() == "0"
+                            ? character.HitPoints
+                            : updatedCharacter.HitPoints;
+                    character.Strength =
+                        updatedCharacter.Strength.ToString() == "0"
+                            ? character.Strength
+                            : updatedCharacter.Strength;
+                    character.Defense =
+                        updatedCharacter.Defense.ToString() == "0"
+                            ? character.Defense
+                            : updatedCharacter.Defense;
+                    character.Intelligence =
+                        updatedCharacter.Intelligence.ToString() == "0"
+                            ? character.Intelligence
+                            : updatedCharacter.Intelligence;
+                    character.Class =
+                        updatedCharacter.Class.ToString() == "0"
+                            ? character.Class
+                            : updatedCharacter.Class;
 
-                character.Name =
-                    updatedCharacter.Name == null ? character.Name : updatedCharacter.Name;
-                character.HitPoints =
-                    updatedCharacter.HitPoints.ToString() == "0"
-                        ? character.HitPoints
-                        : updatedCharacter.HitPoints;
-                character.Strength =
-                    updatedCharacter.Strength.ToString() == "0"
-                        ? character.Strength
-                        : updatedCharacter.Strength;
-                character.Defense =
-                    updatedCharacter.Defense.ToString() == "0"
-                        ? character.Defense
-                        : updatedCharacter.Defense;
-                character.Intelligence =
-                    updatedCharacter.Intelligence.ToString() == "0"
-                        ? character.Intelligence
-                        : updatedCharacter.Intelligence;
-                character.Class =
-                    updatedCharacter.Class.ToString() == "0"
-                        ? character.Class
-                        : updatedCharacter.Class;
+                    await _context.SaveChangesAsync();
 
-                await _context.SaveChangesAsync();
-
-                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                    serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found";
+                }
             }
             catch (Exception ex)
             {
@@ -123,6 +160,44 @@ namespace dotnet_rpg.Services.CharacterService
             }
 
             return serviceResponse;
+        }
+
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharacterSkill)
+        {
+            var response = new ServiceResponse<GetCharacterDto>();
+            try
+            {
+                var character = await _context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .FirstOrDefaultAsync(c => c.Id == newCharacterSkill.CharacterId &&
+                c.User.Id == GetUserId());
+
+                if(character == null)
+                {
+                    response.Success = false;
+                    response.Message = "Character not found.";
+                    return response;
+                }
+
+                var skill = await _context.Skills.FirstOrDefaultAsync( s => s.Id == newCharacterSkill.SkillId);
+                if(skill == null)
+                {
+                    response.Success = false;
+                    response.Message = "Skill not found";
+                    return response;
+                }
+
+                character.Skills.Add(skill);
+                await _context.SaveChangesAsync();
+                response.Data = _mapper.Map<GetCharacterDto>(character);
+            }
+            catch (System.Exception ex)
+            {
+                response.Success = false;
+                response.Message = ex.Message;
+            }
+            return response;
         }
     }
 }
